Fix issues with empty tabix data for region, label display, and page loading errors
